feat/step4 입력 view 수정
