+++ src/main/java/step3and4/ui/InputView.java
@@ -5,14 +5,19 @@ import java.util.Scanner;
 
 public class InputView {
     private static final int MINIMUM_NUMBER = 0;
+    private static final Scanner SCANNER = new Scanner(System.in);
 
-    public static InputRequestByStep3 createInputRequest() {
-        Scanner scanner = new Scanner(System.in);
-        System.out.println("자동차 대수는 몇 대 인가요?");
-        int carNumber = createIntNumber(scanner);
+    public static InputRequest createInputRequest() {
+        System.out.println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).");
+        String[] carNamesSplit = createCarNames();
         System.out.println("시도할 횟수는 몇 회 인가여?");
-        int tryNumber = createIntNumber(scanner);
-        return new InputRequestByStep3(carNumber, tryNumber);
+        int tryNumber = createIntNumber(SCANNER);
+        return new InputRequest(carNamesSplit, tryNumber);
+    }
+
+    private static String[] createCarNames() {
+        String carNames = SCANNER.next();
+        return carNames.split(",");
     }
 
     private static int createIntNumber(Scanner scanner) {
